Structure code into classes
