+++ src/pyalgebralib/config.py
@@ -1,0 +1,18 @@
+symbols = [
+    "0",
+    "1",
+    "2",
+    "3",
+    "4",
+    "5",
+    "6",
+    "7",
+    "8",
+    "9",
+    "a",
+    "b",
+    "c",
+    "d",
+    "e",
+    "f",
+]  # Symbols list
